pyproject.toml
[tool.poetry]
name = "sevinch-med"
version = "0.1.0"
description = ""
authors = ["zukhriddin <zux.828@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
gunicorn = "^23.0.0"
django = "^5.1.1"
dj-database-url = "^2.2.0"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
sentry-sdk = {extras = ["django"], version = "^2.14.0"}
django-nested-admin = "^4.1.1"
python-dateutil = "^2.9.0.post0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
